ai-service/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn>=0.30.0,<1.0
sentence-transformers>=3.0.0,<4.0
faiss-cpu>=1.9.0,<2.0
sqlalchemy>=2.0.30,<3.0
aiosqlite>=0.20.0,<1.0
httpx>=0.27.0,<1.0
python-dotenv>=1.0.1,<2.0
pydantic>=2.8.0,<3.0
sse-starlette>=2.0.0,<3.0